Use BCL CipherMode enum for AesCipher class just like TripleDesCipher class; Create a dedicated AesCtrCipher class just like AesGcmCipher class.
