+++ src/Renci.SshNet/PrivateKeyFile.SSHCOM.cs
@@ -9,8 +9,6 @@ using Renci.SshNet.Common;
 using Renci.SshNet.Security;
 using Renci.SshNet.Security.Cryptography.Ciphers;
 
-using CipherMode = System.Security.Cryptography.CipherMode;
-
 namespace Renci.SshNet
 {
     public partial class PrivateKeyFile
@@ -53,7 +51,7 @@ namespace Renci.SshNet
                     }
 
                     var key = GetCipherKey(_passPhrase, 192 / 8);
-                    using var ssh2Сipher = new TripleDesCipher(key, new byte[8], CipherMode.CBC, pkcs7Padding: false);
+                    using var ssh2Сipher = new TripleDesCipher(key, new byte[8], CipherMode.CBC);
                     keyData = ssh2Сipher.Decrypt(dataReader.ReadBytes(blobSize));
                 }
                 else
