+++ test/Renci.SshNet.Benchmarks/Security/Cryptography/Ciphers/AesCipherBenchmarks.cs
@@ -1,4 +1,6 @@
-﻿using BenchmarkDotNet.Attributes;
+﻿using System.Security.Cryptography;
+
+using BenchmarkDotNet.Attributes;
 
 using Renci.SshNet.Security.Cryptography.Ciphers;
 
@@ -26,49 +28,37 @@ namespace Renci.SshNet.Benchmarks.Security.Cryptography.Ciphers
         [Benchmark]
         public byte[] Encrypt_CBC()
         {
-            return new AesCipher(_key, _iv, AesCipherMode.CBC, false).Encrypt(_data);
+            return new AesCipher(_key, _iv, CipherMode.CBC, false).Encrypt(_data);
         }
 
         [Benchmark]
         public byte[] Decrypt_CBC()
         {
-            return new AesCipher(_key, _iv, AesCipherMode.CBC, false).Decrypt(_data);
-        }
-
-        [Benchmark]
-        public byte[] Encrypt_CFB()
-        {
-            return new AesCipher(_key, _iv, AesCipherMode.CFB, false).Encrypt(_data);
-        }
-
-        [Benchmark]
-        public byte[] Decrypt_CFB()
-        {
-            return new AesCipher(_key, _iv, AesCipherMode.CFB, false).Decrypt(_data);
+            return new AesCipher(_key, _iv, CipherMode.CBC, false).Decrypt(_data);
         }
 
         [Benchmark]
         public byte[] Encrypt_CTR()
         {
-            return new AesCipher(_key, _iv, AesCipherMode.CTR, false).Encrypt(_data);
+            return new AesCtrCipher(_key, _iv).Encrypt(_data);
         }
 
         [Benchmark]
         public byte[] Decrypt_CTR()
         {
-            return new AesCipher(_key, _iv, AesCipherMode.CTR, false).Decrypt(_data);
+            return new AesCtrCipher(_key, _iv).Decrypt(_data);
         }
 
         [Benchmark]
         public byte[] Encrypt_ECB()
         {
-            return new AesCipher(_key, null, AesCipherMode.ECB, false).Encrypt(_data);
+            return new AesCipher(_key, null, CipherMode.ECB, false).Encrypt(_data);
         }
 
         [Benchmark]
         public byte[] Decrypt_ECB()
         {
-            return new AesCipher(_key, null, AesCipherMode.ECB, false).Decrypt(_data);
+            return new AesCipher(_key, null, CipherMode.ECB, false).Decrypt(_data);
         }
     }
 }
