+++ test/Renci.SshNet.Benchmarks/Security/Cryptography/Ciphers/TripleDesCipherBenchmarks.cs
@@ -1,8 +1,8 @@
-﻿using BenchmarkDotNet.Attributes;
+﻿using System.Security.Cryptography;
+
+using BenchmarkDotNet.Attributes;
 
 using Renci.SshNet.Security.Cryptography.Ciphers;
-
-using CipherMode = System.Security.Cryptography.CipherMode;
 
 namespace Renci.SshNet.Benchmarks.Security.Cryptography.Ciphers
 {
@@ -35,18 +35,6 @@ namespace Renci.SshNet.Benchmarks.Security.Cryptography.Ciphers
         public byte[] Decrypt_CBC()
         {
             return new TripleDesCipher(_key, _iv, CipherMode.CBC, false).Decrypt(_data);
-        }
-
-        [Benchmark]
-        public byte[] Encrypt_CFB()
-        {
-            return new TripleDesCipher(_key, _iv, CipherMode.CFB, false).Encrypt(_data);
-        }
-
-        [Benchmark]
-        public byte[] Decrypt_CFB()
-        {
-            return new TripleDesCipher(_key, _iv, CipherMode.CFB, false).Decrypt(_data);
         }
     }
 }
